Implemented Eslint to these files and some minor changes
- implemented eslint
- adjusted some routes with put, delete, and get (:id) methods to check the existence of the data

+++ src/routes/parent-router.ts
@@ -1,5 +1,5 @@
-import { Hono } from "hono"
-import { prisma } from "../../prisma/db.js"
+import { Hono } from 'hono'
+import { prisma } from '../../prisma/db.js'
 // import { Parent } from "../db/schemas/parent-schema"
 
 const parentRouter = new Hono()
@@ -17,11 +17,26 @@ parentRouter
     const parent = await prisma.parents.findUnique({
       where: {
         id: Number.parseInt(id),
-      }
+      },
     })
 
-    return c.json({ parent })
+    const students = await prisma.students.findMany({
+      where: {
+        parents_students: {
+          some: {
+            parent_id: Number.parseInt(id),
+          },
+        },
+      },
+    })
+
+    if (!parent) {
+      return c.json({ message: 'Parent not found' }, 400)
+    }
+
+    return c.json({ parent, students })
   })
+
   // POST /api/parents
   .post('/', async (c) => {
     const {
@@ -31,32 +46,45 @@ parentRouter
       pekerjaan,
       tahun_lahir,
       penghasilan,
-    } = await c.req.parseBody() as Record<string, string>
+      email
+    } = await c.req.json()
 
-    await prisma.parents.create({
+    const parent = await prisma.parents.create({
       data: {
         nama,
         jenjang_pendidikan,
         NIK,
         pekerjaan,
+        email,
         penghasilan,
         tahun_lahir: Number.parseInt(tahun_lahir),
-      }
+      },
     })
 
-    return c.json({ success: true })
+    return c.json({ message: 'success', parent })
   })
 
   // DELETE /api/parents/:id
   .delete('/:id', async (c) => {
     const id = c.req.param('id')
-    const parent = await prisma.parents.delete({
+
+    const parent = await prisma.parents.findFirst({
       where: {
         id: Number.parseInt(id),
-      }
+      },
     })
 
-    return c.json({ parent })
+    if (!parent) {
+      return c.json({ message: 'Parent not found' }, 400)
+    }
+
+    await prisma.parents.delete({
+      where: {
+        id: Number.parseInt(id),
+      },
+    })
+
+    return c.json({ message: 'success', parent })
   })
 
   // PUT /api/parents/:id
@@ -68,10 +96,23 @@ parentRouter
       NIK,
       pekerjaan,
       tahun_lahir,
-      penggayahan,
-    } = await c.req.parseBody() as Record<string, string>
+      penghasilan,
+      email
+    } = await c.req.json()
 
-    const parent = await prisma.parents.update({
+    const parent = await prisma.parents.findFirst({
+      where: {
+        id: Number.parseInt(id),
+      },
+    })
+
+    if (!parent) {
+      return c.json({
+        message: 'Parent not found',
+      }, 400)
+    }
+
+    await prisma.parents.update({
       where: {
         id: Number.parseInt(id),
       },
@@ -79,12 +120,14 @@ parentRouter
         nama,
         jenjang_pendidikan,
         NIK,
+        email,
         pekerjaan,
         tahun_lahir: Number.parseInt(tahun_lahir),
-      }
+        penghasilan,
+      },
     })
 
-    return c.json({ parent })
+    return c.json({ message: 'success', parent })
   })
 
 export default parentRouter
